\name{anscombe}
\alias{anscombe}
\docType{data}
\title{Anscombe's data on graphics}
\description{
Four regression problems, y1|x1, y2|x1,
y3|x1 and y4|x2 that illustrate the importance of graphics.
}
\format{This data frame contains the following columns:
 \describe{
    \item{x1}{
 Predictor with response y1, y2 or y3
    }
    \item{y1}{
Response 1 
    }
    \item{y2}{
Response 2 
    }
    \item{y3}{
Response 3 
    }
    \item{x2}{
 Predictor with response y4
    }
    \item{y4}{
Response 4 
    }
 }
}
\details{}
\source{
F. Anscombe (1983).  Graphs in statistical analysis, \emph{American
Statistician}, 27, 17-21.}
\references{Weisberg, S. (2005).  \emph{Applied Linear Regression}, 3rd edition.
New York: Wiley, Section 1.4}
\examples{
data(anscombe)
}
\keyword{datasets}
