\name{cathedral}
\alias{cathedral}
\docType{data}
\non_function{}
\title{Cathedrals}
\description{
Heights and lengths of Gothic and Romanesque cathedrals.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Type}{
Romanesque or Gothic
    }
    \item{Height}{
Total height, feet 
    }
    \item{Length}{
Total length, feet 
    }
    \item{Cathedral}{
Name of the Cathedral
    }
 }
}

\source{Stephen Jay Gould}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 6.7.}
\examples{
data(cathedral)
}
\keyword{datasets}
