\name{longley}
\alias{longley}
\non_function{}
\docType{data}
\title{Longley data}
\description{
Longley (1967) data to demonstrate inadequacies of regression computer programs then available.
}
\format{This data frame contains the following columns of national US data:
 \describe{
    \item{GNP}{
 GNP(millions of dollars)
    }
    \item{Def}{
 GNP deflator (percent)
    }
    \item{Unemp}{
 Unemployed (thousands)
    }
    \item{Mil}{
 Armed force size (thousands)
    }
    \item{Pop}{
 Population (thousands)
    }
    \item{Year}{
 Year 
    }
    \item{Emp}{
 Number employed (thousands)
    }
 }
}

\source{
Longley, J. (1967).  An appraisal of least squares programs for the
electronic computer from the point of view of the user.  \emph{J. Amer. Statist.
Assoc.}, 62, 819-41.}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 4.10.}
\examples{
data(longley)
}
\keyword{datasets}
