\name{UN3}
\alias{UN3}
\docType{data}
\title{ National statistics from the United Nations }
\description{
National health, welfare, and education statistics for 127
places, mostly UN members, but also other areas like Hong Kong that are not
independent countries.  Localities with missing data have been removed from this
data set.
}
\format{
  This data frame uses locality name as row labels.
  In some cases, the geographic area is
    smaller than a country; for example, Hong Kong.
It contains the following columns:
  \describe{
    \item{ModernC}{Percent of unmarried women using a modern method of
    contraception}
    \item{Change}{Annual population growth rate, percent}
    \item{PPgdp}{Per capita 2001 GDP, in US \$}
    \item{Frate}{Percent of females over age 15 economically active}
    \item{Pop}{Population, thousands}
    \item{Fertility}{Expected number of live births per female, 2000}
    \item{Purban}{Percent of population that is urban, 2001}
  }
}
\details{
 These data were collected at published by the UN from a variety of sources.
 See original source for additional footnotes concerning values for
 individual countries.  Locality names are given in the first column of the
 data file.
}
\source{
 \url{http://unstats.un.org/unsd/demographic}
}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 7.7.
}
\examples{
head(UN3)
}
\keyword{datasets}
